# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: TAXI IMPULSE API
servers:
  - url: /api
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/register-driver:
    post:
      operationId: registerDriver
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDriverRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterDriverResponse'
...
  /drivers:
    get:
      operationId: getDrivers
      tags: [drivers]
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: city
          schema:
            type: string
        - in: query
          name: isApproved
          schema:
            type: boolean
        - in: query
          name: pendingApproval
          schema:
            type: boolean
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Driver'
    post:
      operationId: createDriver
      tags: [drivers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDriverRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Driver'
  /drivers/{id}:
    get:
      operationId: getDriverById
      tags: [drivers]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Driver'
    patch:
      operationId: updateDriver
      tags: [drivers]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDriverRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Driver'
  /drivers/{id}/status:
    patch:
      operationId: updateDriverStatus
      tags: [drivers]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DriverStatusRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Driver'
  /drivers/{id}/block:
    patch:
      operationId: blockDriver
      tags: [drivers]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlockDriverRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Driver'
  /drivers/{id}/approve:
    patch:
      operationId: approveDriver
      tags: [drivers]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveDriverRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Driver'
...
  /chat/{orderId}/messages:
    get:
      operationId: getChatMessages
      tags: [chat]
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatMessage'
    post:
      operationId: sendChatMessage
      tags: [chat]
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessage'
  /support/messages:
    get:
      operationId: getSupportMessages
      tags: [support]
      parameters:
        - in: query
          name: userId
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SupportMessage'
    post:
      operationId: sendSupportMessage
      tags: [support]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendSupportMessageRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportMessage'
components:
  schemas:
...
    LoginRequest:
      type: object
      properties:
        phone:
          type: string
        password:
          type: string
      required: [phone, password]
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        password:
          type: string
        city:
          type: string
      required: [name, phone, password]
    RegisterDriverRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        password:
          type: string
        city:
          type: string
        workCity:
          type: string
        carModel:
          type: string
        carColor:
          type: string
        carNumber:
          type: string
        licenseNumber:
          type: string
        experience:
          type: integer
      required: [name, phone, password, carModel, carNumber]
    RegisterDriverResponse:
      type: object
      properties:
        message:
          type: string
        userId:
          type: integer
      required: [message, userId]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/User'
      required: [token, user]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [passenger, driver, admin]
        city:
          type: string
          nullable: true
        rating:
          type: number
          nullable: true
        totalRides:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
          nullable: true
      required: [id, name, phone, role]
...
    Driver:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        phone:
          type: string
        city:
          type: string
        workCity:
          type: string
          nullable: true
        carModel:
          type: string
        carColor:
          type: string
          nullable: true
        carNumber:
          type: string
        licenseNumber:
          type: string
          nullable: true
        experience:
          type: integer
          nullable: true
        status:
          type: string
          enum: [online, offline, busy]
        rating:
          type: number
          nullable: true
        totalRides:
          type: integer
          nullable: true
        balance:
          type: number
          nullable: true
        tariffId:
          type: integer
          nullable: true
        tariffName:
          type: string
          nullable: true
        isApproved:
          type: boolean
          nullable: true
        isBlocked:
          type: boolean
          nullable: true
        autoAssign:
          type: boolean
          nullable: true
        rejectionReason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
          nullable: true
      required: [id, userId, name, phone, city, carModel, carNumber, status]
    CreateDriverRequest:
      type: object
      properties:
        userId:
          type: integer
        carModel:
          type: string
        carColor:
          type: string
        carNumber:
          type: string
        city:
          type: string
        workCity:
          type: string
        tariffId:
          type: integer
        licenseNumber:
          type: string
        experience:
          type: integer
      required: [userId, carModel, carNumber, city]
    UpdateDriverRequest:
      type: object
      properties:
        carModel:
          type: string
        carColor:
          type: string
        carNumber:
          type: string
        city:
          type: string
        workCity:
          type: string
        tariffId:
          type: integer
        isApproved:
          type: boolean
        isBlocked:
          type: boolean
        autoAssign:
          type: boolean
        balance:
          type: number
        licenseNumber:
          type: string
        experience:
          type: integer
        rejectionReason:
          type: string
    DriverStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [online, offline, busy]
      required: [status]
    BlockDriverRequest:
      type: object
      properties:
        isBlocked:
          type: boolean
      required: [isBlocked]
    ApproveDriverRequest:
      type: object
      properties:
        isApproved:
          type: boolean
        rejectionReason:
          type: string
      required: [isApproved]
...
    ChatMessage:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        senderId:
          type: integer
        senderName:
          type: string
        senderRole:
          type: string
          enum: [passenger, driver]
        message:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, orderId, senderId, senderName, senderRole, message, createdAt]
    SendMessageRequest:
      type: object
      properties:
        senderId:
          type: integer
        senderRole:
          type: string
          enum: [passenger, driver]
        message:
          type: string
      required: [senderId, senderRole, message]
    SupportMessage:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userRole:
          type: string
        message:
          type: string
        isFromSupport:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, userId, userRole, message, isFromSupport, createdAt]
    SendSupportMessageRequest:
      type: object
      properties:
        userId:
          type: integer
        userRole:
          type: string
        message:
          type: string
        isFromSupport:
          type: boolean
      required: [userId, userRole, message]